requirements.txt
PySide6~=6.7.2
pillow~=10.4.0
PyQt6~=6.7.1
roman~=4.2